Added User Authentication using Email and Google Auth through Passport js under issue #16
User can now choose to login with both Email and Password and using Google Account.

To check the endpoints for simple login, send Post requests to the specified api end points ("/register", "/login", "check-session") using terminal or postman with the required JSON payload.

To check the google auth functionality go to the /auth/google endpoint on running the server.

The stored collections for both the login methods is same.

All the necessary credentials are saved in my .env while for security reasons. Ask me for necessary credentials if you want to test the functionality or add your own in your .env file.